clean up

+++ lib/core/network/constants.h
@@ -5,8 +5,8 @@ namespace core
 {
     constexpr uint8_t PAYLOAD_MINIMUM_LENGTH = 8;
 
-    constexpr uint8_t BROADCAST_ID = 0;
-    constexpr uint8_t SERVER_ID = 1;
+    constexpr uint32_t BROADCAST_ID = 0;
+    constexpr uint32_t SERVER_ID = 1;
 }
 
 #endif // P2P_TECHDEMO_LIB_CORE_NETWORK_CONSTANTS_H_
